[core] Functorize Http(s) Client signatures
This introduces the Response functor, that sort of sets an interface for
implementations of the Client API.

It includes a refactor of the current `Httpkit_lwt.Client.Https` to use
the functor, and a draft of `Httpkit_lwt.Client.Http`.

Because they both share the same signature, they should be
interchangeable 🚀

In the future I think the Client module should just be split into
Request and Response, and that we should be building a datatype for the
Request that can be realized through some of the implementations
(similarly to how the Server.Middleware stack works).

Thanks @ulrikstrid for the inspiration 🙌🏼